% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckl_names.R
\docType{data}
\name{ckl_names}
\alias{ckl_names}
\title{Italian Checklist of Vascular Flora: names}
\format{
ckl_names is a dataframe with 26113 rows and 3 columns:
\describe{
\item{codice_unico}{unique identifier of the plant name}
\item{entita}{accepted plant name}
\item{sinonimo}{synonym}
}
}
\source{
\url{https://www.dryades.units/floritaly}
}
\description{
A dataset containing the complete list of vascular plant names - accepted and synonyms -
recorded in Italy
}
