% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckl_version.R
\docType{data}
\name{ckl_version}
\alias{ckl_version}
\title{Italian Checklist of Vascular Flora: version id}
\format{
ckl_version is a dataframe with 1 row and 1 column:
\describe{
\item{version}{alphanumeric, version of Italian Checklist of Vascular Flora}
}
}
\source{
\url{https://www.dryades.units/floritaly}
}
\description{
A dataframe containing the columns "version" and a single row
}
