% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckl_version.R
\docType{data}
\name{ckl_version}
\alias{ckl_version}
\title{Italian Checklist of Vascular Flora: version}
\format{
ckl_version is a dataframe with 1 row and 1 columns:
\describe{
\item{version}{version of the Checklist, in the format yyyy.n, where yyyy is the year of release, n is the issue}
}

ckl_version is a dataframe with 1 row and 1 column.
ckl_parsed is a dataframe with parsed names from ckl_names. Parsing is done via rgnparser
}
\source{
\url{https://dryades.units.it/api_test/floritaly/version}
}
\description{
The version of the Checklist of the Italian vascular flora.
}
