% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckl_version.R
\docType{data}
\name{ckl_version}
\alias{ckl_version}
\title{Italian Checklist of Vascular Flora: version}
\format{
ckl_version is a dataframe with 1 row and 3 columns:
\describe{
\item{major}{major number of the package version}
\item{minor}{minor number of the package version}
\item{patch}{patch number of the package version}
}
}
\source{
\url{https://dryades.units.it/api_test/floritaly/version}
}
\description{
The version of the Checklist of the Italian vascular flora.
}
