% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superparse.R
\name{superparse}
\alias{superparse}
\title{Super parser for scientific names}
\usage{
superparse(namevek)
}
\arguments{
\item{namevek}{a character vector containing scientific names}
}
\value{
a dataframe with many columns
}
\description{
Extracts generic, specific, and infraspecific epithets along with
authorship, after parsing with rngparser::gn_parse_tidy
}
\examples{
my_names <- c("Crocus etruscus Parl.","Santolina pinnata Viv.")
s <- superparse(my_names)

}
